# lib/api-zod/src/generated/types/authToken.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Birla Open Minds International School API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AuthToken {
  token: string;
  admin: AdminUser;
}