# lib/api-zod/src/generated/types/facultyMember.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Birla Open Minds International School API
 * OpenAPI spec version: 0.1.0
 */

export interface FacultyMember {
  id: number;
  name: string;
  subject: string;
  qualification: string;
  /** @nullable */
  photoUrl?: string | null;
  sortOrder?: number;
  createdAt: string;
}