# lib/api-zod/src/generated/types/galleryItemInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Birla Open Minds International School API
 * OpenAPI spec version: 0.1.0
 */
import type { GalleryItemInputCategory } from './galleryItemInputCategory';

export interface GalleryItemInput {
  /** @minLength 1 */
  title: string;
  imageUrl: string;
  category: GalleryItemInputCategory;
}