# lib/api-zod/src/generated/types/loginCredentials.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Birla Open Minds International School API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginCredentials {
  username: string;
  password: string;
}