# lib/api-zod/src/generated/types/newsEvent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Birla Open Minds International School API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsEventType } from './newsEventType';

export interface NewsEvent {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  type: NewsEventType;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  eventDate?: string | null;
  publishedAt: string;
  createdAt: string;
}